perf: refactor enemy and player into component based organization
