+++ Cargo.toml
@@ -9,4 +9,5 @@ edition = "2021"
 anyhow = "1.0.70"
 bevy = "0.10.1"
 bevy_editor_pls = "0.4.0"
+bevy_rapier2d = "0.21.0"
 rand = "0.8.5"
