--- src/components/mod.rs
@@ -1,2 +1,0 @@
-pub mod enemy;
-pub mod player;
