+++ src/enemy/mod.rs
@@ -1,0 +1,5 @@
+mod components;
+mod resources;
+mod systems;
+
+pub mod plugins;
