+++ src/player/resources.rs
@@ -1,0 +1,1 @@
+
